refactored form, added button, fixed typo
