+++ src/components/Form/form.js
@@ -1,8 +1,9 @@
 import React, {useState} from "react"
-import UploadAndDisplayImage from 'components/ImageUplaod/imageUpload';
+import UploadAndDisplayImage from 'components/ImageUpload/imageUpload';
+import AddMoreFieldsButton from "components/AddMoreFields/addMoreFields";
 
 const Form = () => {
-
+    
     return (
         <form>
             Properties of your NFT
@@ -13,7 +14,11 @@ const Form = () => {
             <label htmlFor='nftdescription'> Description</label>
             <input type="text" name="nftdescription" id="nftdescription"></input>
             <br></br>
-            <UploadAndDisplayImage/>
+            <UploadAndDisplayImage />
+            <br></br>
+            <AddMoreFieldsButton />
+            <br></br>
+            <input type="submit" value="Mint NFT"></input>
         </form>
     )
 }
